Move buffer, image, format to base. Move mesh and texture to new asset module. Swap scoped_image out by new raw_image class. Add CMake option LAVA_RESOURCE_LOADING (default ON). Turn off to disable build of resource module and selene. Load icon from .c file to remove app depenmdency to resource functions. Update DOCS.md
