+++ liblava/app/app_icon.c
@@ -1,0 +1,68 @@
+// Converted using https://lvgl.io/tools/imageconverter
+// Pixel format:  Red: 8 bit, Green: 8 bit, Blue: 8 bit, Alpha: 8 bit
+const unsigned char app_icon[] = {
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x44, 0x00, 0x00, 0x00, 0xc8, 0x00, 0x00, 0x00, 0xc8, 0x00, 0x00, 0x00, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x00, 0x00, 0xac, 0x02, 0x02, 0x02, 0xff, 0x33, 0x33, 0x33, 0xff, 0x33, 0x33, 0x33, 0xff, 0x02, 0x02, 0x02, 0xff, 0x00, 0x00, 0x00, 0xac, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0f, 0x00, 0x00, 0x00, 0x84, 0x00, 0x00, 0x00, 0xf0, 0x13, 0x13, 0x13, 0xff, 0x94, 0x94, 0x94, 0xff, 0xf2, 0xf2, 0xf2, 0xff, 0xf2, 0xf2, 0xf2, 0xff, 0x94, 0x94, 0x94, 0xff, 0x13, 0x13, 0x13, 0xff, 0x00, 0x00, 0x00, 0xf0, 0x00, 0x00, 0x00, 0x84, 0x00, 0x00, 0x00, 0x0f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x00, 0xe7, 0x0b, 0x0b, 0x0b, 0xff, 0x6f, 0x6f, 0x6f, 0xff, 0xec, 0xec, 0xec, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xec, 0xec, 0xec, 0xff, 0x6f, 0x6f, 0x6f, 0xff, 0x0b, 0x0b, 0x0b, 0xff, 0x00, 0x00, 0x00, 0xe7, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x37, 0x00, 0x00, 0x00, 0xc4, 0x00, 0x00, 0x00, 0xfc, 0x48, 0x48, 0x48, 0xff, 0xca, 0xca, 0xca, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xca, 0xca, 0xca, 0xff, 0x48, 0x48, 0x48, 0xff, 0x00, 0x00, 0x00, 0xfc, 0x00, 0x00, 0x00, 0xc4, 0x00, 0x00, 0x00, 0x37, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x23, 0x00, 0x00, 0x00, 0xa7, 0x00, 0x00, 0x00, 0xfc, 0x2f, 0x2f, 0x2f, 0xff, 0xaf, 0xaf, 0xaf, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xaf, 0xaf, 0xaf, 0xff, 0x2f, 0x2f, 0x2f, 0xff, 0x00, 0x00, 0x00, 0xfc, 0x00, 0x00, 0x00, 0xa7, 0x00, 0x00, 0x00, 0x23, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x7f, 0x00, 0x00, 0x00, 0xeb, 0x10, 0x10, 0x10, 0xff, 0x88, 0x88, 0x88, 0xff, 0xf1, 0xf1, 0xf1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf1, 0xf1, 0xf1, 0xff, 0x88, 0x88, 0x88, 0xff, 0x10, 0x10, 0x10, 0xff, 0x00, 0x00, 0x00, 0xeb, 0x00, 0x00, 0x00, 0x7f, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x5c, 0x00, 0x00, 0x00, 0xdc, 0x0a, 0x0b, 0x0b, 0xff, 0x63, 0x63, 0x63, 0xff, 0xe8, 0xe8, 0xe8, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe8, 0xe8, 0xe8, 0xff, 0x63, 0x63, 0x63, 0xff, 0x0a, 0x0b, 0x0b, 0xff, 0x00, 0x00, 0x00, 0xdc, 0x00, 0x00, 0x00, 0x5c, 0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x2f, 0x00, 0x00, 0x00, 0xb7, 0x00, 0x00, 0x00, 0xff, 0x39, 0x39, 0x39, 0xff, 0xc7, 0xc7, 0xc7, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xc7, 0xc7, 0xc7, 0xff, 0x39, 0x39, 0x39, 0xff, 0x00, 0x00, 0x00, 0xff, 0x00, 0x00, 0x00, 0xb7, 0x00, 0x00, 0x00, 0x2f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x97, 0x00, 0x00, 0x00, 0xff, 0x25, 0x25, 0x25, 0xff, 0xaa, 0xaa, 0xaa, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xaa, 0xaa, 0xaa, 0xff, 0x25, 0x25, 0x25, 0xff, 0x00, 0x00, 0x00, 0xff, 0x00, 0x00, 0x00, 0x97, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x6f, 0x00, 0x00, 0x00, 0xe8, 0x09, 0x09, 0x09, 0xff, 0x82, 0x82, 0x82, 0xff, 0xec, 0xec, 0xec, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xec, 0xec, 0xec, 0xff, 0x82, 0x82, 0x82, 0xff, 0x09, 0x09, 0x09, 0xff, 0x00, 0x00, 0x00, 0xe8, 0x00, 0x00, 0x00, 0x6f, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x4c, 0x00, 0x00, 0x00, 0xd8, 0x06, 0x06, 0x06, 0xff, 0x5e, 0x5e, 0x5e, 0xff, 0xde, 0xde, 0xde, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xec, 0xeb, 0xec, 0xff, 0x80, 0x80, 0x80, 0xff, 0x80, 0x80, 0x80, 0xff, 0xec, 0xeb, 0xec, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xde, 0xde, 0xde, 0xff, 0x5e, 0x5e, 0x5e, 0xff, 0x06, 0x06, 0x06, 0xff, 0x00, 0x00, 0x00, 0xd8, 0x00, 0x00, 0x00, 0x4c, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x27, 0x00, 0x00, 0x00, 0xb4, 0x00, 0x00, 0x00, 0xfb, 0x32, 0x32, 0x32, 0xff, 0xb9, 0xb9, 0xb9, 0xff, 0xfe, 0xfe, 0xfe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xa9, 0xa9, 0xa9, 0xff, 0x24, 0x24, 0x24, 0xff, 0x02, 0x01, 0x01, 0xff, 0x02, 0x01, 0x01, 0xff, 0x24, 0x24, 0x24, 0xff, 0xa9, 0xa9, 0xa9, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xfe, 0xfe, 0xff, 0xb9, 0xb9, 0xb9, 0xff, 0x32, 0x32, 0x32, 0xff, 0x00, 0x00, 0x00, 0xfb, 0x00, 0x00, 0x00, 0xb4, 0x00, 0x00, 0x00, 0x27, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x90, 0x00, 0x00, 0x00, 0xf8, 0x22, 0x22, 0x22, 0xff, 0x9b, 0x9b, 0x9b, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xc7, 0xc7, 0xc7, 0xff, 0x39, 0x39, 0x39, 0xff, 0x02, 0x01, 0x01, 0xff, 0x3e, 0x07, 0x0b, 0xff, 0xad, 0x0f, 0x1d, 0xff, 0xad, 0x0f, 0x1d, 0xff, 0x3e, 0x07, 0x0b, 0xff, 0x02, 0x01, 0x01, 0xff, 0x39, 0x39, 0x39, 0xff, 0xc7, 0xc7, 0xc7, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x9b, 0x9b, 0x9b, 0xff, 0x22, 0x22, 0x22, 0xff, 0x00, 0x00, 0x00, 0xf8, 0x00, 0x00, 0x00, 0x90, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x6b, 0x00, 0x00, 0x00, 0xe0, 0x09, 0x09, 0x09, 0xff, 0x71, 0x71, 0x71, 0xff, 0xe9, 0xe9, 0xe9, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe8, 0xe8, 0xe8, 0xff, 0x62, 0x62, 0x62, 0xff, 0x0a, 0x0a, 0x0a, 0xff, 0x1f, 0x04, 0x06, 0xff, 0x8b, 0x0d, 0x17, 0xff, 0xcc, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcc, 0x10, 0x20, 0xff, 0x8b, 0x0d, 0x17, 0xff, 0x1f, 0x04, 0x06, 0xff, 0x0a, 0x0a, 0x0a, 0xff, 0x62, 0x62, 0x62, 0xff, 0xe8, 0xe8, 0xe8, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe9, 0xe9, 0xe9, 0xff, 0x71, 0x71, 0x71, 0xff, 0x09, 0x09, 0x09, 0xff, 0x00, 0x00, 0x00, 0xe0, 0x00, 0x00, 0x00, 0x6b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0xbc, 0x00, 0x00, 0x00, 0xff, 0x40, 0x40, 0x40, 0xff, 0xdb, 0xdb, 0xdb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf1, 0xf1, 0xf1, 0xff, 0x87, 0x88, 0x87, 0xff, 0x0f, 0x0f, 0x0f, 0xff, 0x13, 0x03, 0x04, 0xff, 0x6d, 0x0b, 0x13, 0xff, 0xcc, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcc, 0x10, 0x20, 0xff, 0x6d, 0x0b, 0x13, 0xff, 0x13, 0x03, 0x04, 0xff, 0x0f, 0x0f, 0x0f, 0xff, 0x87, 0x88, 0x87, 0xff, 0xf1, 0xf1, 0xf1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xdb, 0xdb, 0xdb, 0xff, 0x4d, 0x4d, 0x4d, 0xff, 0x03, 0x03, 0x03, 0xff, 0x00, 0x00, 0x00, 0xbc, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0xff, 0x20, 0x20, 0x20, 0xff, 0x9f, 0x9f, 0x9f, 0xff, 0xfb, 0xfb, 0xfb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xae, 0xae, 0xae, 0xff, 0x2e, 0x2e, 0x2e, 0xff, 0x03, 0x01, 0x01, 0xff, 0x4e, 0x08, 0x0f, 0xff, 0xb6, 0x0f, 0x1d, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xb6, 0x0f, 0x1d, 0xff, 0x4e, 0x08, 0x0f, 0xff, 0x03, 0x01, 0x01, 0xff, 0x2e, 0x2e, 0x2e, 0xff, 0xae, 0xae, 0xae, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe5, 0xe5, 0xe5, 0xff, 0x59, 0x59, 0x59, 0xff, 0x03, 0x03, 0x03, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x62, 0x62, 0x62, 0xff, 0x52, 0x52, 0x52, 0xff, 0x04, 0x04, 0x04, 0xff, 0x32, 0x32, 0x32, 0xff, 0xc1, 0xc0, 0xc1, 0xff, 0xfb, 0xfb, 0xfb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xc9, 0xca, 0xc9, 0xff, 0x47, 0x47, 0x47, 0xff, 0x03, 0x01, 0x01, 0xff, 0x32, 0x06, 0x0a, 0xff, 0xa7, 0x0f, 0x1c, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xa7, 0x0f, 0x1c, 0xff, 0x32, 0x06, 0x0a, 0xff, 0x03, 0x01, 0x01, 0xff, 0x47, 0x47, 0x47, 0xff, 0xc9, 0xca, 0xc9, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xea, 0xea, 0xea, 0xff, 0x7b, 0x7b, 0x7b, 0xff, 0x0a, 0x0a, 0x0a, 0xff, 0x1a, 0x1a, 0x1a, 0xff, 0x35, 0x35, 0x35, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xfc, 0xfc, 0xfc, 0xff, 0xaf, 0xaf, 0xaf, 0xff, 0x2c, 0x2c, 0x2c, 0xff, 0x08, 0x08, 0x08, 0xff, 0x59, 0x59, 0x59, 0xff, 0xe2, 0xe2, 0xe2, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xec, 0xec, 0xec, 0xff, 0x6f, 0x6f, 0x6f, 0xff, 0x0b, 0x0b, 0x0b, 0xff, 0x16, 0x04, 0x05, 0xff, 0x87, 0x0d, 0x17, 0xff, 0xc9, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xc9, 0x10, 0x20, 0xff, 0x87, 0x0d, 0x17, 0xff, 0x16, 0x04, 0x05, 0xff, 0x0b, 0x0b, 0x0b, 0xff, 0x6f, 0x6f, 0x6f, 0xff, 0xec, 0xec, 0xec, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x9a, 0x9a, 0x9a, 0xff, 0x25, 0x25, 0x25, 0xff, 0x07, 0x07, 0x07, 0xff, 0x67, 0x67, 0x67, 0xff, 0xe4, 0xe4, 0xe4, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0x8c, 0x8d, 0x8c, 0xff, 0x1b, 0x1b, 0x1b, 0xff, 0x0c, 0x0c, 0x0c, 0xff, 0x7d, 0x7d, 0x7d, 0xff, 0xed, 0xed, 0xed, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf2, 0xf2, 0xf2, 0xff, 0x94, 0x94, 0x94, 0xff, 0x13, 0x13, 0x13, 0xff, 0x0e, 0x02, 0x03, 0xff, 0x69, 0x0b, 0x12, 0xff, 0xc6, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xc6, 0x10, 0x20, 0xff, 0x69, 0x0b, 0x12, 0xff, 0x0e, 0x02, 0x03, 0xff, 0x13, 0x13, 0x13, 0xff, 0x94, 0x94, 0x94, 0xff, 0xf2, 0xf2, 0xf2, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xb5, 0xb5, 0xb5, 0xff, 0x32, 0x32, 0x32, 0xff, 0x05, 0x05, 0x05, 0xff, 0x4d, 0x4d, 0x4d, 0xff, 0xd5, 0xd5, 0xd5, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe3, 0xe3, 0xe3, 0xff, 0x68, 0x68, 0x68, 0xff, 0x05, 0x05, 0x05, 0xff, 0x28, 0x28, 0x28, 0xff, 0x89, 0x89, 0x89, 0xff, 0x31, 0x31, 0x31, 0xff, 0x01, 0x02, 0x01, 0xff, 0x48, 0x08, 0x0e, 0xff, 0xae, 0x0f, 0x1d, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xae, 0x0f, 0x1d, 0xff, 0x48, 0x08, 0x0e, 0xff, 0x01, 0x02, 0x01, 0xff, 0x31, 0x31, 0x31, 0xff, 0xa7, 0xa7, 0xa7, 0xff, 0x54, 0x54, 0x54, 0xff, 0x05, 0x05, 0x05, 0xff, 0x2d, 0x2d, 0x2d, 0xff, 0xb7, 0xb7, 0xb7, 0xff, 0xfa, 0xfa, 0xfa, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd3, 0xd3, 0xd3, 0xff, 0x42, 0x42, 0x42, 0xff, 0x01, 0x01, 0x01, 0xff, 0x02, 0x01, 0x01, 0xff, 0x6f, 0x0c, 0x14, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd0, 0x10, 0x20, 0xff, 0x87, 0x0e, 0x18, 0xff, 0x08, 0x02, 0x02, 0xff, 0x00, 0x00, 0x00, 0xff, 0x1f, 0x1f, 0x1f, 0xff, 0x9a, 0x9a, 0x9a, 0xff, 0xfa, 0xfa, 0xfa, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x08, 0x02, 0x02, 0xff, 0x07, 0x01, 0x02, 0xff, 0x54, 0x09, 0x0f, 0xff, 0xbc, 0x10, 0x1e, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcb, 0x10, 0x20, 0xff, 0x79, 0x0c, 0x15, 0xff, 0x15, 0x03, 0x04, 0xff, 0x00, 0x00, 0x00, 0xff, 0x04, 0x04, 0x04, 0xff, 0xc5, 0xc5, 0xc5, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x8a, 0x0c, 0x16, 0xff, 0x71, 0x0b, 0x14, 0xff, 0x12, 0x02, 0x04, 0xff, 0x0b, 0x02, 0x03, 0xff, 0x72, 0x0b, 0x13, 0xff, 0xc3, 0x10, 0x1f, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcb, 0x10, 0x20, 0xff, 0x90, 0x0d, 0x18, 0xff, 0x1f, 0x04, 0x06, 0xff, 0x06, 0x01, 0x01, 0xff, 0x4d, 0x08, 0x0e, 0xff, 0x76, 0x0b, 0x14, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xb4, 0x0f, 0x1d, 0xff, 0x52, 0x09, 0x0f, 0xff, 0x00, 0x00, 0x01, 0xff, 0x23, 0x04, 0x07, 0xff, 0x95, 0x0e, 0x19, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xab, 0x0f, 0x1c, 0xff, 0x38, 0x07, 0x0b, 0xff, 0x04, 0x01, 0x01, 0xff, 0x30, 0x06, 0x0a, 0xff, 0x9c, 0x0e, 0x1a, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xa5, 0x0f, 0x1c, 0xff, 0x38, 0x06, 0x0a, 0xff, 0x01, 0x01, 0x01, 0xff, 0x36, 0x07, 0x0b, 0xff, 0xa9, 0x0f, 0x1c, 0xff, 0xce, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xb5, 0x0f, 0x1d, 0xff, 0x4f, 0x09, 0x0f, 0xff, 0x03, 0x01, 0x01, 0xff, 0x23, 0x04, 0x07, 0xff, 0x87, 0x0d, 0x17, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xca, 0x10, 0x20, 0xff, 0x84, 0x0c, 0x17, 0xff, 0x1a, 0x04, 0x06, 0xff, 0x0b, 0x02, 0x02, 0xff, 0x58, 0x09, 0x10, 0xff, 0xc3, 0x10, 0x1f, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xca, 0x11, 0x20, 0xff, 0x69, 0x0b, 0x12, 0xff, 0x11, 0x03, 0x04, 0xff, 0x0c, 0x02, 0x03, 0xff, 0x6d, 0x0b, 0x13, 0xff, 0xc3, 0x10, 0x1f, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xc9, 0x11, 0x20, 0xff, 0x66, 0x0a, 0x12, 0xff, 0x0f, 0x02, 0x03, 0xff, 0x10, 0x03, 0x04, 0xff, 0x77, 0x0c, 0x15, 0xff, 0xc7, 0x10, 0x1f, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xc9, 0x10, 0x20, 0xff, 0x82, 0x0c, 0x16, 0xff, 0x1a, 0x04, 0x06, 0xff, 0x0a, 0x01, 0x02, 0xff, 0x54, 0x09, 0x0f, 0xff, 0xbf, 0x10, 0x1f, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xb2, 0x0f, 0x1d, 0xff, 0x46, 0x08, 0x0d, 0xff, 0x02, 0x01, 0x01, 0xff, 0x2c, 0x05, 0x09, 0xff, 0x97, 0x0e, 0x1a, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd0, 0x10, 0x20, 0xff, 0x9f, 0x0f, 0x1b, 0xff, 0x34, 0x06, 0x0a, 0xff, 0x01, 0x01, 0x01, 0xff, 0x37, 0x07, 0x0b, 0xff, 0xa8, 0x0f, 0x1b, 0xff, 0xce, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xa1, 0x0f, 0x1b, 0xff, 0x2d, 0x05, 0x08, 0xff, 0x02, 0x00, 0x01, 0xff, 0x42, 0x08, 0x0d, 0xff, 0xaa, 0x0f, 0x1c, 0xff, 0xac, 0x0f, 0x1c, 0xff, 0x47, 0x08, 0x0d, 0xff, 0x01, 0x01, 0x00, 0xff, 0x27, 0x05, 0x08, 0xff, 0x98, 0x0e, 0x1a, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xc8, 0x10, 0x1f, 0xff, 0x80, 0x0c, 0x16, 0xff, 0x13, 0x03, 0x04, 0xff, 0x0c, 0x02, 0x03, 0xff, 0x0d, 0x02, 0x03, 0xff, 0x11, 0x03, 0x04, 0xff, 0x7b, 0x0c, 0x15, 0xff, 0xc6, 0x10, 0x1f, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xc5, 0x10, 0x1f, 0xff, 0x29, 0x04, 0x07, 0xff, 0x28, 0x04, 0x07, 0xff, 0xc4, 0x10, 0x1f, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0x05, 0x05, 0x05, 0xff, 0x90, 0x0b, 0x16, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x90, 0x0b, 0x16, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xc9, 0xc9, 0xc9, 0xff, 0x05, 0x05, 0x05, 0xff, 0x8f, 0x0b, 0x17, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0x8f, 0x0b, 0x17, 0xff, 0x05, 0x05, 0x05, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xac, 0xac, 0xac, 0xff, 0x27, 0x27, 0x27, 0xff, 0x00, 0x00, 0x00, 0xff, 0x2a, 0x05, 0x08, 0xff, 0x9d, 0x0f, 0x1a, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd0, 0x10, 0x20, 0xff, 0x9d, 0x0f, 0x1a, 0xff, 0x2a, 0x05, 0x08, 0xff, 0x01, 0x01, 0x01, 0xff, 0x6c, 0x6c, 0x6c, 0xff, 0xf6, 0xf6, 0xf6, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfc, 0xfc, 0xfc, 0xff, 0xc6, 0xc6, 0xc6, 0xff, 0x39, 0x39, 0x39, 0xff, 0x02, 0x02, 0x02, 0xff, 0x49, 0x49, 0x49, 0xff, 0xa2, 0xa2, 0xa2, 0xff, 0x31, 0x31, 0x31, 0xff, 0x01, 0x02, 0x01, 0xff, 0x48, 0x08, 0x0e, 0xff, 0xae, 0x0f, 0x1d, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xae, 0x0f, 0x1d, 0xff, 0x48, 0x08, 0x0e, 0xff, 0x01, 0x02, 0x01, 0xff, 0x31, 0x31, 0x31, 0xff, 0x67, 0x67, 0x67, 0xff, 0x0e, 0x0e, 0x0e, 0xff, 0x1f, 0x1f, 0x1f, 0xff, 0xa6, 0xa6, 0xa6, 0xff, 0xf9, 0xf9, 0xf9, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe5, 0xe5, 0xe5, 0xff, 0x5f, 0x5f, 0x5f, 0xff, 0x0a, 0x0a, 0x0a, 0xff, 0x26, 0x26, 0x26, 0xff, 0xa6, 0xa6, 0xa6, 0xff, 0xfa, 0xfa, 0xfa, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf2, 0xf2, 0xf2, 0xff, 0x94, 0x94, 0x94, 0xff, 0x13, 0x13, 0x13, 0xff, 0x0e, 0x02, 0x03, 0xff, 0x69, 0x0b, 0x12, 0xff, 0xc6, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xc6, 0x10, 0x20, 0xff, 0x69, 0x0b, 0x12, 0xff, 0x0e, 0x02, 0x03, 0xff, 0x13, 0x13, 0x13, 0xff, 0x94, 0x94, 0x94, 0xff, 0xf2, 0xf2, 0xf2, 0xff, 0xfe, 0xfe, 0xfe, 0xff, 0xcc, 0xcc, 0xcc, 0xff, 0x40, 0x40, 0x40, 0xff, 0x02, 0x02, 0x02, 0xff, 0x4c, 0x4c, 0x4d, 0xff, 0xd3, 0xd3, 0xd3, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x86, 0x86, 0x86, 0xff, 0xed, 0xed, 0xed, 0xff, 0x80, 0x80, 0x80, 0xff, 0x0c, 0x0d, 0x0c, 0xff, 0x18, 0x18, 0x18, 0xff, 0x87, 0x87, 0x87, 0xff, 0xfa, 0xfa, 0xfa, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xec, 0xec, 0xec, 0xff, 0x6f, 0x6f, 0x6f, 0xff, 0x0b, 0x0b, 0x0b, 0xff, 0x16, 0x04, 0x05, 0xff, 0x87, 0x0d, 0x17, 0xff, 0xc9, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xc9, 0x10, 0x20, 0xff, 0x87, 0x0d, 0x17, 0xff, 0x16, 0x04, 0x05, 0xff, 0x0b, 0x0b, 0x0b, 0xff, 0x6f, 0x6f, 0x6f, 0xff, 0xeb, 0xeb, 0xeb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xaa, 0xab, 0xaa, 0xff, 0x21, 0x21, 0x21, 0xff, 0x02, 0x02, 0x02, 0xff, 0x78, 0x78, 0x78, 0xff, 0xe9, 0xe9, 0xe9, 0xff, 0xff, 0xff, 0xff, 0xff, 0x86, 0x86, 0x86, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x43, 0x43, 0x43, 0xff, 0x29, 0x29, 0x29, 0xff, 0x05, 0x05, 0x05, 0xff, 0x65, 0x65, 0x65, 0xff, 0xe1, 0xe1, 0xe1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xc9, 0xca, 0xc9, 0xff, 0x47, 0x47, 0x47, 0xff, 0x03, 0x01, 0x01, 0xff, 0x32, 0x06, 0x0a, 0xff, 0xa7, 0x0f, 0x1c, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd0, 0x10, 0x20, 0xff, 0xa7, 0x0f, 0x1c, 0xff, 0x32, 0x06, 0x0a, 0xff, 0x03, 0x01, 0x01, 0xff, 0x47, 0x47, 0x47, 0xff, 0xc9, 0xca, 0xc9, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe8, 0xe8, 0xe8, 0xff, 0x7a, 0x7a, 0x7a, 0xff, 0x04, 0x04, 0x04, 0xff, 0x21, 0x21, 0x21, 0xff, 0xa7, 0xa7, 0xa7, 0xff, 0x84, 0x84, 0x84, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0xdf, 0x01, 0x01, 0x01, 0xff, 0x41, 0x41, 0x41, 0xff, 0xd1, 0xd1, 0xd1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xae, 0xae, 0xae, 0xff, 0x2e, 0x2e, 0x2e, 0xff, 0x03, 0x01, 0x01, 0xff, 0x4e, 0x08, 0x0f, 0xff, 0xb6, 0x0f, 0x1d, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xb6, 0x0f, 0x1d, 0xff, 0x4e, 0x08, 0x0f, 0xff, 0x03, 0x01, 0x01, 0xff, 0x2e, 0x2e, 0x2e, 0xff, 0xae, 0xae, 0xae, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd7, 0xd7, 0xd7, 0xff, 0x4d, 0x4d, 0x4d, 0xff, 0x02, 0x02, 0x02, 0xff, 0x0d, 0x0d, 0x0d, 0xff, 0x00, 0x00, 0x00, 0xdf, 0x00, 0x00, 0x00, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0xbc, 0x01, 0x01, 0x01, 0xff, 0x4b, 0x4b, 0x4b, 0xff, 0xdb, 0xdb, 0xdb, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf1, 0xf1, 0xf1, 0xff, 0x87, 0x88, 0x87, 0xff, 0x0f, 0x0f, 0x0f, 0xff, 0x13, 0x03, 0x04, 0xff, 0x6d, 0x0b, 0x13, 0xff, 0xcc, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xd1, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xd1, 0x10, 0x20, 0xff, 0xcf, 0x10, 0x20, 0xff, 0xcc, 0x10, 0x20, 0xff, 0x6d, 0x0b, 0x13, 0xff, 0x13, 0x03, 0x04, 0xff, 0x0f, 0x0f, 0x0f, 0xff, 0x87, 0x88, 0x87, 0xff, 0xf1, 0xf1, 0xf1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xd6, 0xd6, 0xd6, 0xff, 0x22, 0x22, 0x22, 0xff, 0x00, 0x00, 0x00, 0xff, 0x00, 0x00, 0x00, 0xbc, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x6b, 0x00, 0x00, 0x00, 0xe0, 0x09, 0x09, 0x09, 0xff, 0x71, 0x71, 0x71, 0xff, 0xe9, 0xe9, 0xe9, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe8, 0xe8, 0xe8, 0xff, 0x62, 0x62, 0x62, 0xff, 0x0a, 0x0a, 0x0a, 0xff, 0x1f, 0x04, 0x06, 0xff, 0x8b, 0x0d, 0x17, 0xff, 0xce, 0x10, 0x20, 0xff, 0x3a, 0x05, 0x09, 0xff, 0x3a, 0x05, 0x09, 0xff, 0xce, 0x10, 0x20, 0xff, 0x8b, 0x0d, 0x17, 0xff, 0x1f, 0x04, 0x06, 0xff, 0x0a, 0x0a, 0x0a, 0xff, 0x62, 0x62, 0x62, 0xff, 0xe8, 0xe8, 0xe8, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe9, 0xe9, 0xe9, 0xff, 0x71, 0x71, 0x71, 0xff, 0x09, 0x09, 0x09, 0xff, 0x00, 0x00, 0x00, 0xe0, 0x00, 0x00, 0x00, 0x6b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x90, 0x00, 0x00, 0x00, 0xf8, 0x22, 0x22, 0x22, 0xff, 0x9b, 0x9b, 0x9b, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xc7, 0xc7, 0xc7, 0xff, 0x39, 0x39, 0x39, 0xff, 0x02, 0x01, 0x01, 0xff, 0x3f, 0x07, 0x0c, 0xff, 0x24, 0x04, 0x06, 0xff, 0x24, 0x04, 0x06, 0xff, 0x3f, 0x07, 0x0c, 0xff, 0x02, 0x01, 0x01, 0xff, 0x39, 0x39, 0x39, 0xff, 0xc7, 0xc7, 0xc7, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0x9b, 0x9b, 0x9b, 0xff, 0x22, 0x22, 0x22, 0xff, 0x00, 0x00, 0x00, 0xf8, 0x00, 0x00, 0x00, 0x90, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x27, 0x00, 0x00, 0x00, 0xb4, 0x00, 0x00, 0x00, 0xfb, 0x32, 0x32, 0x32, 0xff, 0xb9, 0xb9, 0xb9, 0xff, 0xfe, 0xfe, 0xfe, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xa9, 0xa9, 0xa9, 0xff, 0x24, 0x24, 0x24, 0xff, 0x00, 0x00, 0x00, 0xff, 0x00, 0x00, 0x00, 0xff, 0x24, 0x24, 0x24, 0xff, 0xa9, 0xa9, 0xa9, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfe, 0xfe, 0xfe, 0xff, 0xb9, 0xb9, 0xb9, 0xff, 0x32, 0x32, 0x32, 0xff, 0x00, 0x00, 0x00, 0xfb, 0x00, 0x00, 0x00, 0xb4, 0x00, 0x00, 0x00, 0x27, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x4c, 0x00, 0x00, 0x00, 0xd8, 0x06, 0x06, 0x06, 0xff, 0x5e, 0x5e, 0x5e, 0xff, 0xde, 0xde, 0xde, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xec, 0xeb, 0xec, 0xff, 0x80, 0x80, 0x80, 0xff, 0x80, 0x80, 0x80, 0xff, 0xec, 0xeb, 0xec, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xde, 0xde, 0xde, 0xff, 0x5e, 0x5e, 0x5e, 0xff, 0x06, 0x06, 0x06, 0xff, 0x00, 0x00, 0x00, 0xd8, 0x00, 0x00, 0x00, 0x4c, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x6f, 0x00, 0x00, 0x00, 0xe8, 0x09, 0x09, 0x09, 0xff, 0x82, 0x82, 0x82, 0xff, 0xec, 0xec, 0xec, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xec, 0xec, 0xec, 0xff, 0x82, 0x82, 0x82, 0xff, 0x09, 0x09, 0x09, 0xff, 0x00, 0x00, 0x00, 0xe8, 0x00, 0x00, 0x00, 0x6f, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x97, 0x00, 0x00, 0x00, 0xff, 0x25, 0x25, 0x25, 0xff, 0xaa, 0xaa, 0xaa, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xaa, 0xaa, 0xaa, 0xff, 0x25, 0x25, 0x25, 0xff, 0x00, 0x00, 0x00, 0xff, 0x00, 0x00, 0x00, 0x97, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x2f, 0x00, 0x00, 0x00, 0xb7, 0x00, 0x00, 0x00, 0xff, 0x39, 0x39, 0x39, 0xff, 0xc7, 0xc7, 0xc7, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xfd, 0xfd, 0xfd, 0xff, 0xc7, 0xc7, 0xc7, 0xff, 0x39, 0x39, 0x39, 0xff, 0x00, 0x00, 0x00, 0xff, 0x00, 0x00, 0x00, 0xb7, 0x00, 0x00, 0x00, 0x2f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x5c, 0x00, 0x00, 0x00, 0xdc, 0x0a, 0x0b, 0x0b, 0xff, 0x63, 0x63, 0x63, 0xff, 0xe8, 0xe8, 0xe8, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xe8, 0xe8, 0xe8, 0xff, 0x63, 0x63, 0x63, 0xff, 0x0a, 0x0b, 0x0b, 0xff, 0x00, 0x00, 0x00, 0xdc, 0x00, 0x00, 0x00, 0x5c, 0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x7f, 0x00, 0x00, 0x00, 0xeb, 0x10, 0x10, 0x10, 0xff, 0x88, 0x88, 0x88, 0xff, 0xf1, 0xf1, 0xf1, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xf1, 0xf1, 0xf1, 0xff, 0x88, 0x88, 0x88, 0xff, 0x10, 0x10, 0x10, 0xff, 0x00, 0x00, 0x00, 0xeb, 0x00, 0x00, 0x00, 0x7f, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x23, 0x00, 0x00, 0x00, 0xa7, 0x00, 0x00, 0x00, 0xfc, 0x2f, 0x2f, 0x2f, 0xff, 0xaf, 0xaf, 0xaf, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xaf, 0xaf, 0xaf, 0xff, 0x2f, 0x2f, 0x2f, 0xff, 0x00, 0x00, 0x00, 0xfc, 0x00, 0x00, 0x00, 0xa7, 0x00, 0x00, 0x00, 0x23, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x37, 0x00, 0x00, 0x00, 0xc4, 0x00, 0x00, 0x00, 0xfc, 0x48, 0x48, 0x48, 0xff, 0xca, 0xca, 0xca, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xca, 0xca, 0xca, 0xff, 0x48, 0x48, 0x48, 0xff, 0x00, 0x00, 0x00, 0xfc, 0x00, 0x00, 0x00, 0xc4, 0x00, 0x00, 0x00, 0x37, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x00, 0xe7, 0x0b, 0x0b, 0x0b, 0xff, 0x6f, 0x6f, 0x6f, 0xff, 0xec, 0xec, 0xec, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xec, 0xec, 0xec, 0xff, 0x6f, 0x6f, 0x6f, 0xff, 0x0b, 0x0b, 0x0b, 0xff, 0x00, 0x00, 0x00, 0xe7, 0x00, 0x00, 0x00, 0x60, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0f, 0x00, 0x00, 0x00, 0x84, 0x00, 0x00, 0x00, 0xf0, 0x13, 0x13, 0x13, 0xff, 0x94, 0x94, 0x94, 0xff, 0xf2, 0xf2, 0xf2, 0xff, 0xf2, 0xf2, 0xf2, 0xff, 0x94, 0x94, 0x94, 0xff, 0x13, 0x13, 0x13, 0xff, 0x00, 0x00, 0x00, 0xf0, 0x00, 0x00, 0x00, 0x84, 0x00, 0x00, 0x00, 0x0f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x00, 0x00, 0xac, 0x02, 0x02, 0x02, 0xff, 0x33, 0x33, 0x33, 0xff, 0x33, 0x33, 0x33, 0xff, 0x02, 0x02, 0x02, 0xff, 0x00, 0x00, 0x00, 0xac, 0x00, 0x00, 0x00, 0x2c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x44, 0x00, 0x00, 0x00, 0xc8, 0x00, 0x00, 0x00, 0xc8, 0x00, 0x00, 0x00, 0x44, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 
+  0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x0b, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+};
