+++ liblava/asset.hpp
@@ -1,0 +1,8 @@
+// copyright : Copyright (c) 2018-present, Lava Block OÜ
+// license   : MIT; see accompanying LICENSE file
+
+#pragma once
+
+#include <liblava/asset/raw_image.hpp>
+#include <liblava/asset/mesh.hpp>
+#include <liblava/asset/texture.hpp>
