+++ liblava/base/image.hpp
@@ -1,4 +1,4 @@
-// file      : liblava/resource/image.hpp
+// file      : liblava/base/image.hpp
 // copyright : Copyright (c) 2018-present, Lava Block OÜ
 // license   : MIT; see accompanying LICENSE file
