+++ liblava/resource/textureloader.hpp
@@ -1,0 +1,20 @@
+// file      : liblava/resource/textureloader.hpp
+// copyright : Copyright (c) 2018-present, Lava Block OÜ
+// license   : MIT; see accompanying LICENSE file
+
+#pragma once
+
+#include <liblava/asset/texture.hpp>
+
+namespace lava {
+
+    texture::ptr load_texture(device_ptr device, file_format filename, texture_type type = texture_type::tex_2d);
+
+    inline texture::ptr load_texture(device_ptr device, string_ref filename, VkFormat format = VK_FORMAT_R8G8B8A8_UNORM, texture_type type = texture_type::tex_2d)
+    {
+        return load_texture(device, { filename, format }, type);
+    }
+
+    texture::ptr create_default_texture(device_ptr device, uv2 size);
+
+} // namespace lava
